private_agent/requirements.txt
httpx==0.27.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
